Re-organise test doubles
Move all the test stubs into their own directory
and facor out some repeated code
