+++ polling_stations/apps/data_collection/tests/stubs/stub_duplicatestation.py
@@ -1,0 +1,15 @@
+import os
+from data_collection.tests.stubs import BaseStubCsvStationsJsonDistrictsImporter
+
+
+"""
+Define a stub implementation of json importer we can run tests against
+uses data that will return a duplicate station on
+(council_id, internal_council_id) by design
+"""
+class Command(BaseStubCsvStationsJsonDistrictsImporter):
+
+    srid             = 4326
+    districts_name   = 'test.geojson'
+    stations_name    = 'test.csv'
+    base_folder_path = os.path.join(os.path.dirname(__file__), '../fixtures/duplicate_station')
