+++ polling_stations/apps/data_collection/tests/stubs/stub_jsonimport.py
@@ -1,0 +1,14 @@
+import os
+from data_collection.tests.stubs import BaseStubCsvStationsJsonDistrictsImporter
+
+
+"""
+Define a stub implementation of json importer we can run tests against
+csv and geojson both use srid 4326
+"""
+class Command(BaseStubCsvStationsJsonDistrictsImporter):
+
+    srid             = 4326
+    districts_name   = 'test.geojson'
+    stations_name    = 'test_4326.csv'
+    base_folder_path = os.path.join(os.path.dirname(__file__), '../fixtures/json_importer')
