+++ polling_stations/apps/data_collection/tests/stubs/stub_kmlimport_different_srids.py
@@ -1,29 +1,19 @@
 import os
-from django.contrib.gis.geos import Point, GEOSGeometry
-from data_collection.management.commands import BaseCsvStationsKmlDistrictsImporter
+from django.contrib.gis.geos import Point
+from data_collection.tests.stubs import BaseStubCsvStationsKmlDistrictsImporter
 
 
 """
 Define a stub implementation of kml importer we can run tests against
 kml uses srid 4326, csv uses srid 27700
 """
-class Command(BaseCsvStationsKmlDistrictsImporter):
+class Command(BaseStubCsvStationsKmlDistrictsImporter):
 
     srid             = 27700
     districts_srid   = 4326
-    council_id       = 'X01000000'
     districts_name   = 'test.kml'
     stations_name    = 'test_27700.csv'
-    base_folder_path = os.path.join(os.path.dirname(__file__), 'fixtures/kml_importer')
-
-    def district_record_to_dict(self, record):
-        geojson = record.geom.geojson
-        poly = self.clean_poly(GEOSGeometry(geojson, srid=self.get_srid('districts')))
-        return {
-            'internal_council_id': record['Name'].value,
-            'name'               : record['Name'].value,
-            'area'               : poly
-        }
+    base_folder_path = os.path.join(os.path.dirname(__file__), '../fixtures/kml_importer')
 
     def station_record_to_dict(self, record):
         location = Point(float(record.east), float(record.north), srid=self.get_srid())
